Added new tests to cover whitelisted file config
Included in the updated fixtures is 1 PDF and 1 Image (jpg file) and their accompanying physical. This is because SilverStripe asset store cannot publish non-existent files, which are needed for our tests.

These fixtures allow us to test reindexing of 'whitelisted' file types.

Also tidy ups on code comments was also applied in this commit (where no context or incorrect ordering of comments are used).